+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # BlogMore ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- The FontAwesome metadata is now cached, saving the need to download it
+  every time a site is built.
+
 ## v2.21.0
 
 **Released: 2026-05-10**